Try to fix long recovery button enabledness delay?
Sometimes the Recover Recordings button stays disabled for a minute or more,
and the loading indicator doesn't even seem to work!

I think it might only happen when there's a lot of recording to delete,
but I'm not sure what's causing the delay. I need to figure out how to reproduce it (put a bunch of garbage data into localforage and localstorage?)
And then investigate whether the page is responsive at all during that time and why the loading indicator isn't working.