Remove global line-height declaration
It's not very friendly to typography. There's not much wrapping text in this app, but the "Plug in a MIDI keyboard and play away!" can now wrap, and it looks ugly if the line-height is so low (it looks squished.)

This spaces some things out, but doesn't seem to break too much stuff. I need to work on the info dialog's heading, though.